+++ .scripts/yaml_validate.py
@@ -9,9 +9,7 @@ exitCode = 0
 pykwalify.init_logging(0)
 
 for file in glob.glob("radar/**/*.yaml"):
-    #print(file)
-    #logging.info(file)
-    c = Core(source_file=file, schema_files=["radar-entry.yaml"])
+    c = Core(source_file=file, schema_files=["src/radar_entry.schema.yaml"])
     try:
         c.validate(raise_exception=True)
     except Exception as e:
