--- cloudbuild.yaml
@@ -1,21 +1,0 @@
-steps:
-  - name: 'gcr.io/cloud-builders/npm:current'
-    entrypoint: 'bash'
-    args:
-      - '-c'
-      - |
-          gcloud init && git config credential.helper gcloud \
-          && git fetch --tags \
-          && npm ci \
-          && npm version patch -m "Release %s" \
-          && git push --follow-tags \
-          && npm run package
-    env:
-      - GIT_COMMITTER_NAME=Pipeline Committer
-      - GIT_AUTHOR_NAME=Pipeline Committer
-      - GIT_COMMITTER_EMAIL=pipeline@extendaretail.com
-      - GIT_AUTHOR_EMAIL='pipeline@extendaretail.com
-  - name: 'gcr.io/cloud-builders/docker'
-    args: [ 'build', '-t', 'eu.gcr.io/$PROJECT_ID/tech-radar:$SHORT_SHA', '.' ]
-images:
-  - 'eu.gcr.io/$PROJECT_ID/tech-radar:$SHORT_SHA'
