+++ src/radar_entry.schema.yaml
