+++ src/resources/lib/radar.js
