+++ src/templates/footer.hbs
